refactor: 같이 실행되지 않는 로직 분리
